Move n_ary_s*.h to drake/systems/.
