+++ drake/systems/n_ary_state.h
